# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Platform RS Studio",
    "description": "Remote sensing data management and workflow orchestration platform.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/datasets/{dataset_id}": {
      "get": {
        "tags": [
          "datasets"
        ],
        "summary": "Get Dataset",
        "operationId": "get_dataset_api_v1_datasets__dataset_id__get",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets/upload-session": {
      "post": {
        "tags": [
          "datasets"
        ],
        "summary": "Create Upload Session",
        "operationId": "create_upload_session_api_v1_datasets_upload_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatasetUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadSessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets/confirm": {
      "post": {
        "tags": [
          "datasets"
        ],
        "summary": "Confirm Dataset Upload",
        "operationId": "confirm_dataset_upload_api_v1_datasets_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatasetUploadConfirmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetVersionSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dataset-versions": {
      "get": {
        "tags": [
          "dataset-versions"
        ],
        "summary": "List Dataset Versions",
        "operationId": "list_dataset_versions_api_v1_dataset_versions_get",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dataset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DatasetVersionSummary"
                  },
                  "title": "Response List Dataset Versions Api V1 Dataset Versions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dataset-versions/{dataset_version_id}": {
      "get": {
        "tags": [
          "dataset-versions"
        ],
        "summary": "Get Dataset Version",
        "operationId": "get_dataset_version_api_v1_dataset_versions__dataset_version_id__get",
        "parameters": [
          {
            "name": "dataset_version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Version Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetVersionSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "DatasetKind": {
        "type": "string",
        "enum": [
          "raster",
          "vector",
          "table",
          "artifact"
        ],
        "title": "DatasetKind"
      },
      "DatasetStatus": {
        "type": "string",
        "enum": [
          "uploaded",
          "processing",
          "ready",
          "failed"
        ],
        "title": "DatasetStatus"
      },
      "DatasetSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "kind": {
            "$ref": "#/components/schemas/DatasetKind"
          },
          "status": {
            "$ref": "#/components/schemas/DatasetStatus"
          },
          "bands": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bands"
          },
          "projection": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Projection"
          },
          "footprint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Footprint"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "workspace_id",
          "name",
          "kind",
          "status",
          "updated_at"
        ],
        "title": "DatasetSummary"
      },
      "DatasetUploadConfirmRequest": {
        "properties": {
          "object_key": {
            "type": "string",
            "title": "Object Key"
          },
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id"
          },
          "dataset_name": {
            "type": "string",
            "title": "Dataset Name"
          },
          "kind": {
            "$ref": "#/components/schemas/DatasetKind"
          }
        },
        "type": "object",
        "required": [
          "object_key",
          "workspace_id",
          "dataset_name",
          "kind"
        ],
        "title": "DatasetUploadConfirmRequest"
      },
      "DatasetUploadRequest": {
        "properties": {
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id"
          },
          "dataset_name": {
            "type": "string",
            "title": "Dataset Name"
          },
          "kind": {
            "$ref": "#/components/schemas/DatasetKind"
          },
          "file_name": {
            "type": "string",
            "title": "File Name"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          }
        },
        "type": "object",
        "required": [
          "workspace_id",
          "dataset_name",
          "kind",
          "file_name",
          "content_type",
          "size_bytes"
        ],
        "title": "DatasetUploadRequest"
      },
      "DatasetVersionSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "status": {
            "$ref": "#/components/schemas/DatasetStatus"
          },
          "asset_path": {
            "type": "string",
            "title": "Asset Path"
          },
          "preview_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preview Url"
          },
          "bbox": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bbox"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "dataset_id",
          "version",
          "status",
          "asset_path",
          "created_at"
        ],
        "title": "DatasetVersionSummary"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UploadSessionResponse": {
        "properties": {
          "object_key": {
            "type": "string",
            "title": "Object Key"
          },
          "upload_url": {
            "type": "string",
            "title": "Upload Url"
          },
          "method": {
            "type": "string",
            "title": "Method",
            "default": "PUT"
          },
          "headers": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Headers"
          }
        },
        "type": "object",
        "required": [
          "object_key",
          "upload_url"
        ],
        "title": "UploadSessionResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}